Implement findUsers(overloaded), removeMessage method and implement when a Server recieves a message with a ACK messageType. findUsers: Both return ClientThread object. One method takes in the User object and the other takes in String. Finds user connect to the server and returns it. Returns null if not found.
RemoveMessage:removes message from the messageBuffer. Takes in a Message object.
 returns the index the message was removed in; -1 if message was not found.
Searches based on the sequence number and the sender information (username).
Once the message is found, it is removed from the messageBuffer and returns the index it was removed in.

ACK messageType: when a server recieves a messageType ACK, it sends a ACK message to the sender(original message sender)
and removes the message from the messageBuffer