Implement sendMessage(MessageType, User, User, Object) in Server.java
Use HashTable to keep track of the ip address for each client connected.
Maps the ipaddress with the client's username.

handlMessage(Message) is now handleMessage(Message, InetAddress)
this for when the client is connecting to the server for the first time.

When a user connects for the first time, a new entry is added to the HashTable.
username as the key and the ipaddress as the value

Update when client disconnects from the server. Just remove the entry from the HashTable.