Create Conversation class in Utilities module
This class will be used in Vector for Client class

Every connected client will have their own conversation.
There will be n-1 conversations created when a client connects. -1 is for leaving out your own Client (the user)

Each conversation will have two sequence numbers.
One for the user and the other for the recipient.
Everytime the recipient ACK the message recieved, the user increments the sequence number
The recipient sequence number is kept to keep track of the messages recieved.
If the difference in recipient sequence number is greater than 1, then the user
didn't recieve a message and the previous message needs to be sent again.

The first time the conversation object is created:
The user sequence number is set when the user connects to the server.
The recipient sequence number is set based on what the recipient has register to in the server.