Add attribute BigInteger to User class
This is used as the sequence number.
This object can hold very large number. 2^37 bit size.
The sequence number is generated by using 14 bits and generating a random number using java.util.Random
That means from 2^0 to 2^14 a random number is generated.
The reason it's being generated at random is to have distinct sequence number for each client connected to the server.