Implement broadcastUserList, clean up code and add comments for handling messages (send and recieve)
Implement broadcastUserList
This is used for sending a message to all connected clients/servers
This is whether a client is connecting to the server or disconnecting.
Everyone who is connected needs to know about it.

Added comments about what needs to be in the Message object
handling a message.

Add code to send in connect clients infomation when a new Client
has connected to the server.