Fine tine message handling (SEND, GET, USERS, DISCONNECT, CONNECT)
Start using sequence numbers for when messages are being sent.

Implement whoIsConnected to print out which client are connected.

Start using Conversation class when a message is sent or received.

Create new Conversation objects when a new client has connected to the server.

Handle message type USERS:
check if there is a conversation for the client in the USERS messageType.
If it's a new client, create a new Conversation object and add it to the vector
else, ignore.

Handle message type DISCONNECT:
remove the client from the conversation vector

Handle message type ACK:
increment sequence number (your own, not the recipient) for the message in the conversation being acknowledged.

Handle message type GET:
Add the message to the appropriate conversation and send a ACK message back to the sender.

Send User object (that contains the connecting Client's info and initial sequence number)