Refine ClientMain class to take in only 5 arguments. username serverIP serverPort recipientUsername clientPort
IT's taking clientPort in order to run the Client on the same
machine as the server.
NOTE: this is used to bind the DatagramSocket, which can only have an
IP address connected to a unique port number. In other words client and server must have different ports.
Needless to say it must be different when running two Client chats on the same machine. So three distinct port numbers.
As long the client chat is sending packets to the proper port for the server, it will work.

This wasn't mentioned in the earlier commits, that the chat needs to provide a client side port number.
It used to be able run and test the chat all on the same machine.