Rename NetworkInfo.java to UserNetworkInfo.java
This will contain the IP address, port and User object of the client
connected to the server.

A vector of UserNetworkInfo will be used to keep track of all
the connected clients to the server.

This is a better option to use because, you can get the User object information
as well the IP address and port. Where using the HashTable, you could not.
This necessary when the newly connected clients or disconnected clients updates
are being sent to other servers .

Make changes to the code accordingly (more needs to be done and tested)

Create getClient(User) and removeClient(User) method
to easily find the UserNetworkInfo object and do operations,
rather than making multiple for loops to acquire the object.