Add setMessageType method in Message class and print to console for error checking and debugging.
setMessageType is used ONLY when server is passing on the information. Instead of
making new Mess object, use the existing one in the messageBuffer and modify
the message type.

Adding print to console to keep track of the events being occured and check for
possible errors. Print to console will be removed later.