Create a new method incrementRecipientSequenceNumber() and set recipient sequence number when creating the Conversation object.
incrementRecipientSequenceNumber :
This for keeping track of the sequence number on the client side. This for checking if the messages are being recieved
in order.
Later this will be used for checking if the difference in sequence number recieved (from message) and
stored (in Conversation object) is 1. If it's more than 1 then, a message has been lost. The client needs to send
a GET message with the specific sequence number of the missing message before accepting any message from that client.

set recipient sequence number :
This for keeping track of the messages being recieved. As mentioned above.